Addition of arrays, subtraction of select
I added an array for the car names, so that it's easier to add them to the ComboBoxModel, and it takes out a few lines of code. I also took out the select option out of every combo box because it would mess up the program if the user went backwards, so this is an attempt at trying to fix that. Currently not doing so great